+++ packages/jest-core/src/__tests__/TestScheduler.test.js
@@ -7,7 +7,7 @@
  */
 
 import {SummaryReporter} from '@jest/reporters';
-import {makeProjectConfig} from '@jest/test-utils';
+import {makeGlobalConfig, makeProjectConfig} from '@jest/test-utils';
 import {createTestScheduler} from '../TestScheduler';
 import * as testSchedulerHelper from '../testSchedulerHelper';
 
@@ -37,18 +37,20 @@ beforeEach(() => {
 
 test('config for reporters supports `default`', async () => {
   const undefinedReportersScheduler = await createTestScheduler(
-    {
+    makeGlobalConfig({
       reporters: undefined,
-    },
+    }),
+    {},
     {},
   );
   const numberOfReporters =
     undefinedReportersScheduler._dispatcher._reporters.length;
 
   const stringDefaultReportersScheduler = await createTestScheduler(
-    {
+    makeGlobalConfig({
       reporters: ['default'],
-    },
+    }),
+    {},
     {},
   );
   expect(stringDefaultReportersScheduler._dispatcher._reporters.length).toBe(
@@ -56,9 +58,10 @@ test('config for reporters supports `default`', async () => {
   );
 
   const defaultReportersScheduler = await createTestScheduler(
-    {
+    makeGlobalConfig({
       reporters: [['default', {}]],
-    },
+    }),
+    {},
     {},
   );
   expect(defaultReportersScheduler._dispatcher._reporters.length).toBe(
@@ -66,16 +69,17 @@ test('config for reporters supports `default`', async () => {
   );
 
   const emptyReportersScheduler = await createTestScheduler(
-    {
+    makeGlobalConfig({
       reporters: [],
-    },
+    }),
+    {},
     {},
   );
   expect(emptyReportersScheduler._dispatcher._reporters.length).toBe(0);
 });
 
 test('.addReporter() .removeReporter()', async () => {
-  const scheduler = await createTestScheduler({}, {});
+  const scheduler = await createTestScheduler(makeGlobalConfig(), {}, {});
   const reporter = new SummaryReporter();
   scheduler.addReporter(reporter);
   expect(scheduler._dispatcher._reporters).toContain(reporter);
@@ -84,7 +88,7 @@ test('.addReporter() .removeReporter()', async () => {
 });
 
 test('schedule tests run in parallel per default', async () => {
-  const scheduler = await createTestScheduler({}, {});
+  const scheduler = await createTestScheduler(makeGlobalConfig(), {}, {});
   const test = {
     context: {
       config: makeProjectConfig({
@@ -107,7 +111,7 @@ test('schedule tests run in parallel per default', async () => {
 });
 
 test('schedule tests run in serial if the runner flags them', async () => {
-  const scheduler = await createTestScheduler({}, {});
+  const scheduler = await createTestScheduler(makeGlobalConfig(), {}, {});
   const test = {
     context: {
       config: makeProjectConfig({
@@ -130,7 +134,11 @@ test('schedule tests run in serial if the runner flags them', async () => {
 });
 
 test('should bail after `n` failures', async () => {
-  const scheduler = await createTestScheduler({bail: 2}, {});
+  const scheduler = await createTestScheduler(
+    makeGlobalConfig({bail: 2}),
+    {},
+    {},
+  );
   const test = {
     context: {
       config: makeProjectConfig({
@@ -162,7 +170,11 @@ test('should bail after `n` failures', async () => {
 });
 
 test('should not bail if less than `n` failures', async () => {
-  const scheduler = await createTestScheduler({bail: 2}, {});
+  const scheduler = await createTestScheduler(
+    makeGlobalConfig({bail: 2}),
+    {},
+    {},
+  );
   const test = {
     context: {
       config: makeProjectConfig({
@@ -194,7 +206,7 @@ test('should not bail if less than `n` failures', async () => {
 });
 
 test('should set runInBand to run in serial', async () => {
-  const scheduler = await createTestScheduler({}, {});
+  const scheduler = await createTestScheduler(makeGlobalConfig(), {}, {});
   const test = {
     context: {
       config: makeProjectConfig({
@@ -220,7 +232,7 @@ test('should set runInBand to run in serial', async () => {
 });
 
 test('should set runInBand to not run in serial', async () => {
-  const scheduler = await createTestScheduler({}, {});
+  const scheduler = await createTestScheduler(makeGlobalConfig(), {}, {});
   const test = {
     context: {
       config: makeProjectConfig({